[Refactor] Intro IO Core
