+++ be/src/fs/fs.h
@@ -20,8 +20,8 @@
 #include "fs/credential/cloud_configuration_factory.h"
 #include "fs/encryption.h"
 #include "gen_cpp/PlanNodes_types.h"
-#include "io/input_stream.h"
-#include "io/seekable_input_stream.h"
+#include "io/core/input_stream.h"
+#include "io/core/seekable_input_stream.h"
 #include "runtime/descriptors.h"
 
 namespace starrocks {
