+++ be/src/fs/fs_starlet.cpp
@@ -34,12 +34,12 @@
 #include "fs/encrypt_file.h"
 #include "fs/output_stream_adapter.h"
 #include "gutil/strings/util.h"
-#include "io/input_stream.h"
+#include "io/core/input_stream.h"
+#include "io/core/output_stream.h"
+#include "io/core/seekable_input_stream.h"
+#include "io/core/throttled_output_stream.h"
+#include "io/core/throttled_seekable_input_stream.h"
 #include "io/io_profiler.h"
-#include "io/output_stream.h"
-#include "io/seekable_input_stream.h"
-#include "io/throttled_output_stream.h"
-#include "io/throttled_seekable_input_stream.h"
 #include "service/staros_worker.h"
 #include "storage/lake/filenames.h"
 #include "storage/olap_common.h"
