+++ be/src/io/compressed_input_stream.h
@@ -20,8 +20,8 @@
 #include "base/container/raw_container.h"
 #include "base/string/slice.h"
 #include "common/status.h"
-#include "io/input_stream.h"
-#include "io/seekable_input_stream.h"
+#include "io/core/input_stream.h"
+#include "io/core/seekable_input_stream.h"
 
 namespace starrocks {
 class StreamDecompressor;
