+++ be/src/io/core/input_stream.h
@@ -18,7 +18,7 @@
 #include <string_view>
 
 #include "common/ownership.h"
-#include "io/readable.h"
+#include "io/core/readable.h"
 
 namespace starrocks::io {
