+++ be/src/io/fd_input_stream.h
@@ -14,7 +14,7 @@
 
 #pragma once
 
-#include "io/seekable_input_stream.h"
+#include "io/core/seekable_input_stream.h"
 
 namespace starrocks::io {
