+++ be/src/runtime/stream_load/stream_load_pipe.h
@@ -40,7 +40,7 @@
 
 #include "base/bit/bit_util.h"
 #include "gen_cpp/Types_types.h"
-#include "io/input_stream.h"
+#include "io/core/input_stream.h"
 #include "runtime/message_body_sink.h"
 #include "util/byte_buffer.h"
 #include "util/compression/stream_decompressor.h"
