+++ be/src/storage/rowset/options.h
@@ -35,7 +35,7 @@
 #pragma once
 
 #include "common/config.h"
-#include "io/seekable_input_stream.h"
+#include "io/core/seekable_input_stream.h"
 #include "storage/olap_common.h"
 #include "storage/options.h"
 #include "storage/rowset/page_handle.h"
