+++ be/test/io/array_input_stream_test.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "io/array_input_stream.h"
+#include "io/core/array_input_stream.h"
 
 #include <gtest/gtest.h>
