+++ be/test/io/seekable_input_stream_test.cpp
@@ -17,8 +17,7 @@
 #include "base/testutil/assert.h"
 #include "base/testutil/parallel_test.h"
 #include "common/logging.h"
-#include "fs/encrypt_file.h"
-#include "io/array_input_stream.h"
+#include "io/core/array_input_stream.h"
 
 namespace starrocks::io {
 
@@ -105,23 +104,6 @@ PARALLEL_TEST(SeekableInputStreamTest, test_read_at_fully) {
     ASSERT_EQ(10, *in.position());
 
     ASSERT_ERROR(in.read_at_fully(1, buff, 10));
-}
-
-// NOLINTNEXTLINE
-PARALLEL_TEST(SeekableInputStreamTest, test_encrypted) {
-    TestInputStream in("0123456789", 5);
-    char buff[10];
-    ASSERT_OK(in.read_at_fully(1, buff, 9));
-    ASSERT_EQ(10, *in.position());
-
-    ASSERT_ERROR(in.read_at_fully(1, buff, 10));
-    ASSERT_FALSE(in.is_encrypted());
-    std::unique_ptr<SeekableInputStream> s =
-            std::make_unique<SeekableInputStreamWrapper>(&in, Ownership::kDontTakeOwnership);
-    ASSERT_FALSE(s->is_encrypted());
-
-    EncryptSeekableInputStream encrypted(std::move(s), {EncryptionAlgorithmPB::AES_128, "0000000000000000"});
-    ASSERT_TRUE(encrypted.is_encrypted());
 }
 
 // NOLINTNEXTLINE
